remove DatabaseOperatorImpl class
